requirements/base.txt
Django==1.8.13
Pillow==3.2.0
dj-database-url==0.4.1
psycopg2==2.6.1
pylibmc==1.5.1
pytz==2016.4
raven==5.15.0

# Storage
django-storages==1.4.1
boto==2.40.0
blanc-contentfiles==0.2.3

# Admin theme
blanc-admin-theme==1.0.1

# Thumbnailing
sorl-thumbnail==12.3

# Glitter
django-glitter==0.1.6
django-mptt==0.8.4
django-mptt-admin==0.3.6
six==1.10.0

# Sortable
django-admin-sortable==2.0.15

# Crispy forms
django-crispy-forms==1.6.0

# Donations
stripe==1.33.0
requests==2.10.0
